Implement chunk creation with temporary position solution
In the  method of the ChunkSerializer, add logic to create a new Chunk instance with a temporary position when a valid  is provided. This is a temporary solution to facilitate the creation of Chunk objects and will be refined in future commits. If the provided  is not found, a  is raised.